Prioritise faster paths, by making cost inversely proportional to speed.

+++ src/main/java/mekanism/common/multipart/MultipartTransporter.java
@@ -373,7 +373,7 @@ public class MultipartTransporter extends MultipartTransmitter<IInventory, Inven
 	}
 
 	@Override
-	public int getCost()
+	public double getCost()
 	{
 		return getPart().getCost();
 	}
